Update clean.py agg
The aggregation function now works to include crop damages, and it gets rid of the multiple tornado per state problem
The only thing to keep in mind now is that when doing count functions on tornados that overlap multiple states is to get unique om values.  Not every row is its own distinct tornado now.