extra pages
I put in some extra pages.  One can be where we write in different analyses we find important (I think Sai has been working on this a lot), and another page where we talk about the project and different assumptions we made.  This helps with the issues we've been creating.